LPS-119068 Fix tests.
